devolução

+++ src/main/java/br/com/univille/herbarium/controller/domain/locacao/DadosLocacaoLivro.java
@@ -1,4 +1,4 @@
 package br.com.univille.herbarium.controller.domain.locacao;
 
-public record DadosLocacaoLivro (Integer idAluno, Integer idLivro) {
+public record DadosLocacaoLivro (Long idAluno, Long idLivro) {
 }
